[Build] Cleaned up a lot of warnings and fixed 2010 build

+++ src/tools/map-extractor/mpq_libmpq.cpp
@@ -90,7 +90,7 @@ MPQFile::MPQFile(const char* filename):
             buffer = 0;
             return;
         }
-        buffer = new char[size];
+        buffer = new char[(int)size];
 
         //libmpq_file_getdata
         libmpq__file_read(mpq_a, filenum, (unsigned char*)buffer, size, &transferred);
@@ -106,10 +106,10 @@ size_t MPQFile::read(void* dest, size_t bytes)
 {
     if (eof) { return 0; }
 
-    size_t rpos = pointer + bytes;
+	size_t rpos = (size_t)(pointer + bytes);
     if (rpos > size)
     {
-        bytes = size - pointer;
+        bytes = (size_t)(size - pointer);
         eof = true;
     }
